add responseCode for dubbo

+++ source/extensions/filters/network/dubbo_proxy/router/router_impl.cc
@@ -34,7 +34,7 @@ FilterStatus Router::onMessageDecoded(MessageMetadataSharedPtr metadata, Context
   if (!route_) {
     ENVOY_STREAM_LOG(debug, "dubbo router: no cluster match for interface '{}'", *callbacks_,
                      invocation.serviceName());
-    callbacks_->streamInfo().setResponseCode(Http::Code::NotFound);
+    callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::NotFound));
     callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::NoRouteFound);
     callbacks_->sendLocalReply(AppException(ResponseStatus::ServiceNotFound,
                                             fmt::format("dubbo router: no route for interface '{}'",
@@ -50,7 +50,7 @@ FilterStatus Router::onMessageDecoded(MessageMetadataSharedPtr metadata, Context
   if (!cluster) {
     ENVOY_STREAM_LOG(debug, "dubbo router: unknown cluster '{}'", *callbacks_,
                      route_entry_->clusterName());
-    callbacks_->streamInfo().setResponseCode(Http::Code::NotFound);
+    callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::NotFound));
     callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::NoClusterFound);                 
     callbacks_->sendLocalReply(
         AppException(ResponseStatus::ServerError, fmt::format("dubbo router: unknown cluster '{}'",
@@ -77,7 +77,7 @@ FilterStatus Router::onMessageDecoded(MessageMetadataSharedPtr metadata, Context
 
   auto conn_pool_data = cluster->tcpConnPool(Upstream::ResourcePriority::Default, this);
   if (!conn_pool_data) {
-    callbacks_->streamInfo().setResponseCode(Http::Code::ServiceUnavailable);
+    callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::ServiceUnavailable));
     callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::NoHealthyUpstream);
     callbacks_->sendLocalReply(
         AppException(
@@ -410,7 +410,7 @@ void Router::UpstreamRequest::onResetStream(ConnectionPool::PoolFailureReason re
   // triggers the release of the current stream at the end of the filter's callback.
   switch (reason) {
   case ConnectionPool::PoolFailureReason::Overflow:
-    parent_.callbacks_->streamInfo().setResponseCode(Http::Code::TooManyRequests);
+    parent_.callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::TooManyRequests));
     parent_.callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::UpstreamOverflow); 
     parent_.callbacks_->sendLocalReply(
         AppException(ResponseStatus::ServerError,
@@ -420,7 +420,7 @@ void Router::UpstreamRequest::onResetStream(ConnectionPool::PoolFailureReason re
   case ConnectionPool::PoolFailureReason::LocalConnectionFailure:
     // Should only happen if we closed the connection, due to an error condition, in which case
     // we've already handled any possible downstream response.
-    parent_.callbacks_->streamInfo().setResponseCode(Http::Code::BadGateway);
+    parent_.callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::BadGateway));
     parent_.callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::UpstreamConnectionFailure);    
     parent_.callbacks_->sendLocalReply(
         AppException(ResponseStatus::ServerError,
@@ -429,7 +429,7 @@ void Router::UpstreamRequest::onResetStream(ConnectionPool::PoolFailureReason re
         false);
     break;
   case ConnectionPool::PoolFailureReason::RemoteConnectionFailure:
-    parent_.callbacks_->streamInfo().setResponseCode(Http::Code::BadGateway);
+    parent_.callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::BadGateway));
     parent_.callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::UpstreamConnectionFailure); 
     parent_.callbacks_->sendLocalReply(
         AppException(ResponseStatus::ServerError,
@@ -438,7 +438,7 @@ void Router::UpstreamRequest::onResetStream(ConnectionPool::PoolFailureReason re
         false);
     break;
   case ConnectionPool::PoolFailureReason::Timeout:
-    parent_.callbacks_->streamInfo().setResponseCode(Http::Code::GatewayTimeout);
+    parent_.callbacks_->streamInfo().setResponseCode(static_cast<uint32_t>(Http::Code::GatewayTimeout));
     parent_.callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::UpstreamRequestTimeout); 
     parent_.callbacks_->sendLocalReply(
         AppException(ResponseStatus::ServerError,
